backend/requirements.txt
supabase
google-genai
python-dotenv
flask
flask-cors
